requirements.txt
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonwebtoken==1.0.0
MarkupSafe==2.1.5
pymongo==4.6.3
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
